Refactor matrix input to enforce integer dimensions
Updated `MatrixInputWindowViewModel` to use `int` for rows
and columns instead of `double`, ensuring valid integer
dimensions. Enhanced `NumericUpDown` controls in
`MatrixInputWindow.axaml` with better UI spacing, input
validation, and formatting. Modified the default
constructor in `MatrixInputWindow.axaml.cs` to initialize
with default dimensions and improved `DataContext` binding
to the `ViewModel`. These changes improve usability,
validation, and maintainability.

+++ Editor/ViewModels/MatrixInputWindowViewModel.cs
@@ -5,8 +5,8 @@ namespace Editor;
 public partial class MatrixInputWindowViewModel : ObservableObject
 {
     [ObservableProperty]
-    private double _rows = 1;
+    private int _rows = 1;
 
     [ObservableProperty]
-    private double _columns = 1;
+    private int _columns = 1;
 }
